# source/app/blueprints/rest/openapi.generated.yaml
openapi: 3.1.0
info:
  title: IRIS
  version: 2.1.0
  description: "IRIS REST API \u2014 automatically generated from the Flask backend."
  contact:
    name: DFIR-IRIS
    email: contact@dfir-iris.org
  license:
    name: LGPLv3
    url: https://www.gnu.org/licenses/lgpl-3.0.txt
security:
- ApiKey: []
paths:
...
  /api/v2/alert-clusters/{identifier}/case:
    delete:
      operationId: unlink_case_delete
      summary: Unlink an alert cluster from its case
      tags:
      - AlertClusters
      security:
      - ApiKey: []
      description: "Same behaviour as `DELETE /api/v2/cases/{case_id}/source-cluster`\nbut rooted at the\
        \ cluster URL so the cluster-detail page's\n\"Unlink from case\" menu can hit it without knowing\
        \ the case id.\nRequires the caller to have write access to the target case \u2014 a\nuser who\
        \ can't touch the case shouldn't be able to strip its\nsource-cluster link either."
      parameters:
      - name: identifier
        in: path
        required: true
        schema:
          type: integer
      responses:
        '200':
          description: Success
        '400':
          $ref: '#/components/responses/ApiError'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /api/v2/cases:
    get:
      operationId: get_cases_get
      summary: List cases
      tags:
      - Cases
      security:
      - ApiKey: []
      description: 'Supports the same query parameters as the case list UI: name,

        customer, classification, owner, state, severity, date ranges,

        and a `quick_search` free-text filter used by the context switcher.'
      parameters:
      - name: page
        in: query
        required: false
        schema:
          type: integer
        description: Page number (default 1)
      - name: per_page
        in: query
        required: false
        schema:
          type: integer
        description: Page size (default 10)
      - name: order_by
        in: query
        required: false
        schema:
          type: string
        description: Column to sort by
      - name: sort_dir
        in: query
        required: false
        schema:
          type: string
        description: 'Sort direction: ''asc'' or ''desc'''
      - name: case_ids
        in: query
        required: false
        schema:
          type: string
        description: Comma-separated list of case identifiers to restrict the result to
      - name: case_customer_id
        in: query
        required: false
        schema:
          type: string
        description: Filter by customer identifier
      - name: case_name
        in: query
        required: false
        schema:
          type: string
        description: Filter by case name (substring match)
      - name: case_description
        in: query
        required: false
        schema:
          type: string
        description: Filter by description (substring match)
      - name: case_classification_id
        in: query
        required: false
        schema:
          type: integer
        description: Filter by classification identifier
      - name: case_owner_id
        in: query
        required: false
        schema:
          type: integer
        description: Filter by owner user identifier
      - name: case_opening_user_id
        in: query
        required: false
        schema:
          type: integer
        description: Filter by the user who opened the case
      - name: case_severity_id
        in: query
        required: false
        schema:
          type: integer
        description: Filter by severity identifier
      - name: case_state_id
        in: query
        required: false
        schema:
          type: integer
        description: Filter by state identifier
      - name: case_soc_id
        in: query
        required: false
        schema:
          type: string
        description: Filter by external SOC identifier
      - name: start_open_date
        in: query
        required: false
        schema:
          type: string
          format: date
        description: Restrict to cases opened on or after this date (YYYY-MM-DD)
      - name: end_open_date
        in: query
        required: false
        schema:
          type: string
          format: date
        description: Restrict to cases opened on or before this date (YYYY-MM-DD)
      - name: start_close_date
        in: query
        required: false
        schema:
          type: string
          format: date
        description: Restrict to cases closed on or after this date (YYYY-MM-DD)
      - name: end_close_date
        in: query
        required: false
        schema:
          type: string
          format: date
        description: Restrict to cases closed on or before this date (YYYY-MM-DD)
      - name: is_open
        in: query
        required: false
        schema:
          type: boolean
        description: Restrict to open (true) or closed (false) cases
      - name: quick_search
        in: query
        required: false
        schema:
          type: string
        description: Free-text search across case name, customer name, and case identifier
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  total:
                    type: integer
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/CaseSchemaForAPIV2'
                  last_page:
                    type: integer
                  current_page:
                    type: integer
                  next_page:
                    type:
                    - integer
                    - 'null'
        '400':
          $ref: '#/components/responses/ApiError'
        '403':
          $ref: '#/components/responses/Forbidden'
    post:
      operationId: create_case_post
      summary: Create a case
      tags:
      - Cases
      security:
      - ApiKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CaseSchemaForAPIV2'
      responses:
        '201':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CaseSchemaForAPIV2'
        '400':
          $ref: '#/components/responses/ApiError'
        '403':
          $ref: '#/components/responses/Forbidden'
...
  /api/v2/cases/{identifier}:
    delete:
      operationId: case_routes_delete_delete
      summary: Delete a case
      tags:
      - Cases
      security:
      - ApiKey: []
      parameters:
      - name: identifier
        in: path
        required: true
        schema:
          type: integer
      responses:
        '204':
          $ref: '#/components/responses/Deleted'
        '400':
          $ref: '#/components/responses/ApiError'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    get:
      operationId: case_routes_get_get
      summary: Get a case
      tags:
      - Cases
      security:
      - ApiKey: []
      parameters:
      - name: identifier
        in: path
        required: true
        schema:
          type: integer
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CaseSchemaForAPIV2'
        '400':
          $ref: '#/components/responses/ApiError'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      operationId: rest_v2_cases_update_put
      summary: Update a case
      tags:
      - Cases
      security:
      - ApiKey: []
      parameters:
      - name: identifier
        in: path
        required: true
        schema:
          type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CaseSchemaForAPIV2'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CaseSchemaForAPIV2'
        '400':
          $ref: '#/components/responses/ApiError'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /api/v2/manage/case-objects/asset-types:
    get:
      operationId: list_taxonomy_get
      summary: list_taxonomy
      tags: &id002
      - ManageCaseObjects
      security:
      - ApiKey: []
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  total:
                    type: integer
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/schema_cls'
                  last_page:
                    type: integer
                  current_page:
                    type: integer
                  next_page:
                    type:
                    - integer
                    - 'null'
        '400':
          $ref: '#/components/responses/ApiError'
        '403':
          $ref: '#/components/responses/Forbidden'
    post:
      operationId: create_taxonomy_post
      summary: create_taxonomy
      tags: &id003
      - ManageCaseObjects
      security:
      - ApiKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/schema_cls'
      responses:
        '201':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/schema_cls'
        '400':
          $ref: '#/components/responses/ApiError'
        '403':
          $ref: '#/components/responses/Forbidden'
...
  /api/v2/manage/case-objects/asset-types/{identifier}/icon/{field}:
    post:
      operationId: upload_asset_type_icon_post
      summary: Upload an asset type icon
      tags:
      - ManageCaseObjects
      security:
      - ApiKey: []
      description: '`field` is either `compromised` or `not_compromised`. The file

        must be sent as multipart form-data under the key `file` and pass

        the existing `allowed_file_icon` check (png/svg only). On success

        the new filename is persisted on the column

        `asset_icon_<field>` and the full updated row is returned so the

        frontend can refresh its detail pane in one round-trip.'
      parameters:
      - name: identifier
        in: path
        required: true
        schema:
          type: integer
      - name: field
        in: path
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssetType'
        '400':
          $ref: '#/components/responses/ApiError'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v2/manage/case-objects/case-classifications:
    get:
      operationId: list_taxonomy_get
      summary: list_taxonomy
      tags: *id002
      security:
      - ApiKey: []
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  total:
                    type: integer
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/schema_cls'
                  last_page:
                    type: integer
                  current_page:
                    type: integer
                  next_page:
                    type:
                    - integer
                    - 'null'
        '400':
          $ref: '#/components/responses/ApiError'
        '403':
          $ref: '#/components/responses/Forbidden'
    post:
      operationId: create_taxonomy_post
      summary: create_taxonomy
      tags: *id003
      security:
      - ApiKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/schema_cls'
      responses:
        '201':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/schema_cls'
        '400':
          $ref: '#/components/responses/ApiError'
        '403':
          $ref: '#/components/responses/Forbidden'
...
  /api/v2/manage/case-objects/case-states:
    get:
      operationId: list_taxonomy_get
      summary: list_taxonomy
      tags: *id002
      security:
      - ApiKey: []
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  total:
                    type: integer
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/schema_cls'
                  last_page:
                    type: integer
                  current_page:
                    type: integer
                  next_page:
                    type:
                    - integer
                    - 'null'
        '400':
          $ref: '#/components/responses/ApiError'
        '403':
          $ref: '#/components/responses/Forbidden'
    post:
      operationId: create_taxonomy_post
      summary: create_taxonomy
      tags: *id003
      security:
      - ApiKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/schema_cls'
      responses:
        '201':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/schema_cls'
        '400':
          $ref: '#/components/responses/ApiError'
        '403':
          $ref: '#/components/responses/Forbidden'
...
  /api/v2/manage/case-objects/evidence-types:
    get:
      operationId: list_taxonomy_get
      summary: list_taxonomy
      tags: *id002
      security:
      - ApiKey: []
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  total:
                    type: integer
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/schema_cls'
                  last_page:
                    type: integer
                  current_page:
                    type: integer
                  next_page:
                    type:
                    - integer
                    - 'null'
        '400':
          $ref: '#/components/responses/ApiError'
        '403':
          $ref: '#/components/responses/Forbidden'
    post:
      operationId: create_taxonomy_post
      summary: create_taxonomy
      tags: *id003
      security:
      - ApiKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/schema_cls'
      responses:
        '201':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/schema_cls'
        '400':
          $ref: '#/components/responses/ApiError'
        '403':
          $ref: '#/components/responses/Forbidden'
...
  /api/v2/war-rooms/{war_room_id}/cases/{case_id}:
    delete:
      operationId: detach_case_delete
      summary: Detach a case from a war room
      tags:
      - WarRooms
      security:
      - ApiKey: []
      parameters:
      - name: war_room_id
        in: path
        required: true
        schema:
          type: integer
      - name: case_id
        in: path
        required: true
        schema:
          type: integer
      responses:
        '204':
          $ref: '#/components/responses/Deleted'
        '400':
          $ref: '#/components/responses/ApiError'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    ApiKey:
      type: apiKey
      in: header
      name: Authorization
      description: "Bearer <api-key> \u2014 the key from your profile."
  responses:
    ApiError:
      description: Error
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
              data: {}
            required:
            - message
    Forbidden:
      description: Insufficient permissions to access resource
    NotFound:
      description: Resource not found
    Deleted:
      description: Resource successfully deleted
  schemas:
...
    CaseSchemaForAPIV2:
      type: object
      properties:
        case_name: {}
        case_description: {}
        case_soc_id: {}
        case_customer_id: {}
        case_organisations:
          type: array
          items:
            type: integer
        protagonists:
          type: array
          items:
            type: object
        case_tags:
          type: string
        classification_id: {}
        reviewer_id: {}
        access_level:
          type: integer
        owner:
          $ref: '#/components/schemas/User'
        severity:
          $ref: '#/components/schemas/Severity'
        classification:
          $ref: '#/components/schemas/CaseClassification'
        reviewer:
          $ref: '#/components/schemas/User'
        tags:
          type: array
          items:
            $ref: '#/components/schemas/Tags'
        state:
          $ref: '#/components/schemas/CaseState'
        case_customer:
          $ref: '#/components/schemas/Customer'
        review_status:
          $ref: '#/components/schemas/ReviewStatus'
      required:
      - case_name
      - case_description
      - case_soc_id
      - case_customer_id
      description: Schema for serializing and deserializing Case objects.
...
    Tags:
      type: object
      properties: {}
...